requirements.txt
discord.py==1.7.3
async-timeout
asyncpraw
youtube_dl
python-dotenv